requirements.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
sentence-transformers==3.0.1
bertopic==0.15.0
hdbscan==0.8.33
umap-learn==0.5.4
plotly==5.17.0
seaborn==0.12.2
matplotlib==3.7.2
scipy==1.11.1
torch>=1.9.0
transformers>=4.34.0
huggingface_hub>=0.15.0,<0.26.0
gensim
networkx